Refactor token verifier
