+++ src/Auth/OAuth2/TokenVerifier/Contract/TokenVerifierInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Northwestern\SysDev\SOA\Auth\OAuth2\TokenVerifier\Contract;
+
+use Lcobucci\JWT\UnencryptedToken;
+
+interface TokenVerifierInterface
+{
+    public function parseAndVerify(string $jwt): UnencryptedToken;
+}
